Allow opting out of DNS management
